Start converting UI elements to razor
