--- code/UI/HUD/Chat/ChatBoxEntry.html
@@ -1,5 +1,0 @@
-﻿<link rel="stylesheet" href="ChatBoxEntry.scss" />
-
-<root>
-    <div class="wrapper" @ref="Wrapper" />
-</root>
