--- code/UI/HUD/GameHUD.html
@@ -1,4 +1,0 @@
-<link rel="stylesheet" href="GameHUD.scss" />
-
-<root>
-</root>
