+++ code/UI/HUD/GameHUD.razor
@@ -1,0 +1,12 @@
+@namespace TFS2
+
+<style>
+    GameHUD {
+        width: 100%;
+        height: 100%;
+        z-index: 0;
+    }
+</style>
+
+<root>
+</root>
