--- code/UI/HUD/Killfeed/KillFeed.html
@@ -1,4 +1,0 @@
-﻿<link rel="stylesheet" href="KillFeed.scss" />
-
-<root>
-</root>
