+++ code/UI/HUD/Killfeed/TFKillFeed.razor
@@ -1,0 +1,6 @@
+﻿@using Sandbox.UI;
+@attribute [StyleSheet("/UI/HUD/Killfeed/TFKillFeed.scss")]
+
+@namespace TFS2.UI
+<root>
+</root>
