+++ code/UI/HUD/Killfeed/TFKillFeedEntry.scss
