--- code/UI/HUD/Player/Health.html
@@ -1,6 +1,0 @@
-<link rel="stylesheet" href="Health.scss" />
-
-<root>
-    <div class="playerclass" @ref="PlayerClassIcon" />
-    <HealthCross />
-</root>
