+++ code/UI/HUD/WinPanel.razor
@@ -1,4 +1,8 @@
-﻿<link rel="stylesheet" href="WinPanel.scss" />
+﻿@using Sandbox.UI;
+
+@attribute [StyleSheet("WinPanel.scss")]
+
+@namespace TFS2
 
 <root>
     <div class="container">
